OMG, do I ever want these to be HTTPS URIs rather than SSH
(Also, I want all the file paths to be relative to $HOME, because not
everywhere will my username be ptolemarch. In fact, in many places I
can't even change it. For example in Crostini.)

The problem is that, on boostrapping, I don't yet have my
dotfiles-private with its SSH keys. So pulling all the submodules
(including, for example, dotfiles-private) is impossible.

Pulling dotfiles-private will require a GitHub Personal Access Token:
https://help.github.com/en/articles/creating-a-personal-access-token-for-the-command-line

Once dotfiles-private is pulled, you probably want to save the
username+token in Git:
  cd .yadm-private
  git config credential.helper store
  git fetch  # you'll have to re-enter username+token here