add submodule printc
For some reason, the README for this repo says to do this:
  git clone https://github.com/philFernandez/printc-.git

(Obviously, I've used `yadm submdoule add` rather than `git clone`.)

I don't know what the hyphen at the end of the repo name is for. If I do
the following:
  git clone https://github.com/philFernandez/printc-.git
  git clone https://github.com/philFernandez/printc.git
  diff -r printc printc-
the only diffs I get are those related to the name.

(This is not true when I clone other GitHub repos.)

Indeed, if I navigate (i.e. in a Web browser) to
  https://github.com/philFernandez/printc-
I'm redirected to
  https://github.com/philFernandez/printc

(This is not true when I browse other GitHub repos.)

I can find no such `printc-` repo among philFernandez's list of repos at
  https://github.com/philFernandez?tab=repositories

Kinda weird, but whatever.